% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_table.R
\name{display_table_server}
\alias{display_table_server}
\title{A Shiny Module Server for Displaying a Dataframe as Table in the Frontend}
\usage{
display_table_server(id, df_table, remove_sub = T)
}
\arguments{
\item{id}{shiny id}

\item{df_table}{dataframe to be displayed}

\item{remove_sub}{Logical, remove "_" in the column names of a dataframe if T.}
}
\value{

}
\description{
a Shiny module server for displaying a dataframe as table in the frontend
}
