% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_table.R
\name{display_table_ui}
\alias{display_table_ui}
\title{A Shiny Module UI for Displaying a Dataframe as Table in the Frontend}
\usage{
display_table_ui(id)
}
\arguments{
\item{id}{shiny id}
}
\value{
shiny input
}
\description{
a Shiny module UI for displaying a dataframe as table in the frontend
}
