% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_server}
\alias{download_server}
\title{Shiny Module Server for Downloading Dataframe}
\usage{
download_server(
  id,
  df_table,
  pre_name,
  metal = metal,
  month = time,
  download_item = "sample"
)
}
\arguments{
\item{id}{shiniy id}

\item{df_table}{the dataframe to be downloaded}

\item{pre_name}{prefix of the name of the download object, char}

\item{metal}{metal selected, char}

\item{month}{month selected, char}

\item{download_item}{a name of the download item, i.e., the title of the dataframe, it would be saved as a part of the download object.}
}
\value{

}
\description{
Shiny Module Server for Downloading Dataframe
}
