% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_ui}
\alias{download_ui}
\title{Shiny Module UI for Downloading Dataframe}
\usage{
download_ui(id, download_item = "sample")
}
\arguments{
\item{id}{shiny id}

\item{download_item}{a name of the download item, i.e., the title of the dataframe, and it would be displayed in the frontend.}
}
\value{
shiny input
}
\description{
Shiny Module UI for Downloading Dataframe
}
