% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_display_sample.R
\name{fn_display_sample}
\alias{fn_display_sample}
\title{Function for Convert a Sample Information Dataframe before Displaying It}
\usage{
fn_display_sample(df)
}
\arguments{
\item{df}{a dataframe to be treated}
}
\value{
a converted dataframe
}
\description{
Convert the sample information dataframe in the expression profiling queries
}
