% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_con_tab_server.R
\name{query_con_abs_server}
\alias{query_con_abs_server}
\title{Server End for Querying the Absolution Metal Concentration}
\usage{
query_con_abs_server(df_input = df_abs_con, input, output, session)
}
\arguments{
\item{df_input}{df_abs_con, a dataframe}

\item{input}{shiny input, metal_2, month_2, animal_id_2, measure_tissue_2}

\item{output}{a list, including element of selected_metal, selected_month and id_choices, and the filtered records of concentration}

\item{session}{}
}
\value{
return a list of output
}
\description{
A shiny server, is designed for querying the absolution metal concentration
}
