% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_con_tab_server.R
\name{query_con_creatinine_server}
\alias{query_con_creatinine_server}
\title{Server End for Querying the Creatinine Concentration}
\usage{
query_con_creatinine_server(
  df_input = df_creatinine_con,
  input,
  output,
  session
)
}
\arguments{
\item{df_input}{df_creatinine_con, a dataframe}

\item{input}{shiny input, metal_1, month_1, animal_id_1}

\item{output}{a list, including element of selected_metal, selected_month and id_choices, and the filtered records of concentration}

\item{session}{}
}
\value{
return a list of output
}
\description{
A shiny server, is designed for querying the Creatinine Concentration
}
