% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_con_tab_server.R
\name{query_con_related_server}
\alias{query_con_related_server}
\title{Server End for Querying the Related Metal Concentration}
\usage{
query_con_related_server(df_input = df_related_con, input, output, session)
}
\arguments{
\item{df_input}{df_related_con, a dataframe}

\item{input}{shiny input, metal_3, month_3, animal_id_3, measure_tissue_3}

\item{output}{a list, including element of selected_metal, selected_month and id_choices, and the filtered records of concentration}

\item{session}{}
}
\value{
return a list of output
}
\description{
A shiny server, is designed for querying the related metal concentration
}
