% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_con.R
\name{query_con_server}
\alias{query_con_server}
\title{A Shiny Module Server End for Querying Data Related to Metal Concentrations}
\usage{
query_con_server(id)
}
\arguments{
\item{id}{shiny id}
}
\value{
a list with three items, selected_metal, selected_month, and df_display
}
\description{
a shiny module server end for querying the metal related data and return the querying condition as well as the filtered records
}
