% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_con.R
\name{query_con_ui}
\alias{query_con_ui}
\title{A Shiny Module UI End for Querying Data Related to Metal Concentrations}
\usage{
query_con_ui(
  id,
  dataset_label = "concentration types",
  dataset_choices = vec_type_choices
)
}
\arguments{
\item{id}{shiny id}

\item{dataset_label}{"concentration types"}

\item{dataset_choices}{"creatinine_concentration", "absolute_metal_concentration" or "related_metal_concentration"}
}
\value{
input data would be transport to the server end of the module
}
\description{
A shiny module ui, is designed for display the query option, dependent on which concentration types you selected.
}
