% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_exprs_select_cohorts.R
\name{query_exprs_select_cohorts_ui}
\alias{query_exprs_select_cohorts_ui}
\title{Title}
\usage{
query_exprs_select_cohorts_ui(id, characteristics = "month_post_implantation")
}
\arguments{
\item{id}{a dataframe of the selected sample information, which is used to update the choice list}

\item{characteristics}{characteristics of the cohorts, "month_post_implantation" or "metal_implanted"}
}
\value{

}
\description{
Title
}
