% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_exprs_select_studies.R
\name{query_exprs_select_studies_server}
\alias{query_exprs_select_studies_server}
\title{title query_exprs_select_studies_server}
\usage{
query_exprs_select_studies_server(id)
}
\arguments{
\item{id}{shiny id}
}
\value{
a list with three element, dataframe: selected_sample, selected_exprs, and a vector: selected_library
}
\description{
title query_exprs_select_studies_server
}
