% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_exprs_select_studies.R
\name{query_exprs_select_studies_ui}
\alias{query_exprs_select_studies_ui}
\title{title query_exprs_select_studies_ui}
\usage{
query_exprs_select_studies_ui(id, dataset_label, dataset_choices)
}
\arguments{
\item{id}{shiny id}

\item{dataset_label}{label of the item, dataset_label="Studies"}

\item{dataset_choices}{the name of the transcription studies, e.g. "serum_miRNA", "serum_tRNA"}
}
\value{

}
\description{
title query_exprs_select_studies_ui
}
