% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_weight_cohorts.R
\name{query_weight_cohorts_server}
\alias{query_weight_cohorts_server}
\title{A Shiny Module Server for Querying Body Weight in Rats Cohorts}
\usage{
query_weight_cohorts_server(id, df_data)
}
\arguments{
\item{id}{shiny id}

\item{df_data}{a big dataframe which contains information of body weight}
}
\value{
generate selected cohorts determined by metal and month post implantation, and the corresponding animal_id and weighting time points.
}
\description{
It is a shiny module server designed for querying body weight data in rats cohorts determined via "Metal Implanted" and "Euthanized Month Post-implantation"
}
