% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_weight_cohorts.R
\name{query_weight_cohorts_ui}
\alias{query_weight_cohorts_ui}
\title{A Shiny Module UI for Querying Body Weight in Rats Cohorts}
\usage{
query_weight_cohorts_ui(id, df_data)
}
\arguments{
\item{id}{shiny id}

\item{df_data}{a big dataframe which contains information of body weight}
}
\value{
return shiny input of "metal" and "euthanized month" selected by user
}
\description{
It is a shiny module ui designed for querying body weight data in rats cohorts determined via "Metal Implanted" and "Euthanized Month Post-implantation"
}
