% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_weight_idx.R
\name{query_weight_idx_server}
\alias{query_weight_idx_server}
\title{A Shiny Module Server for Querying Body Weight Records in Custom Rats Cohorts}
\usage{
query_weight_idx_server(id, df_data, metal, month, l_choices)
}
\arguments{
\item{id}{shiny id}

\item{df_data}{a big dataframe which contains information of body weight}

\item{metal}{the selected metal, provided by the previous model query_weight_cohort}

\item{month}{the selected month, provided by the previous model query_weight_cohort}

\item{l_choices}{animal_id or weighting time point, which used to define the custom cohorts together}
}
\value{
return selected records, TRUE or FALSE.
}
\description{
Update the animal ID or weighting time points according to the selected metal and month post implants.

Using selected animal ID or weighting time points as input, to find out records which is satisfied with the query.
}
