% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_weight_idx.R
\name{query_weight_idx_ui}
\alias{query_weight_idx_ui}
\title{A Shiny Module UI for Querying Body Weight Records in Custom Rats Cohorts}
\usage{
query_weight_idx_ui(id, df_data, label, col)
}
\arguments{
\item{id}{shiny id}

\item{df_data}{a big dataframe which contains information of body weight}

\item{label}{"animal id" or "weighting time"}

\item{col}{"animal_ID" or "weeks_post_implantation", used to select a column of the input dataframe}
}
\value{

}
\description{
Using selected animal ID or weighting time points as input, to find out records which is satisfied with the query.
}
