% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_table.R
\name{remove_sub}
\alias{remove_sub}
\title{Remove the "_" in the Column Names of a Dataframe}
\usage{
remove_sub(df)
}
\arguments{
\item{df}{a dataframe to be functioned}
}
\value{
a new dataframe whose column names is updated
}
\description{
a helper function to remove "_" in the column names of a dataframe
}
